Moved BasicMenu config to the same folder. Renamed all menu configs like menu.config.js.
