+++ src/components/Installation/menu.config.js
