+++ src/components/Node/index.js
@@ -25,7 +25,7 @@ import EndorsedOrganizations from './EndorsedOrganizations';
 import EndorsedDatasets from './EndorsedDatasets';
 import Installations from './Installations';
 import Exception404 from '../exception/404';
-import MenuConfig from './MenuConfig';
+import MenuConfig from './menu.config';
 import withContext from '../hoc/withContext';
 import { BreadCrumbs } from '../widgets';
 import { getSubMenu } from '../../api/util/helpers';
